reconfigured subgraph to recognize factory-template relationship between poolfactory and pool
